+++ kotonebot/client/implements/nemu_ipc/__init__.py
@@ -1,10 +1,17 @@
 # ruff: noqa: E402
 
 from .external_renderer_ipc import ExternalRendererIpc
-from .nemu_ipc import NemuIpcImpl, NemuIpcImplConfig
+from .nemu_ipc import (
+    NemuIpcDisplayNotFoundError,
+    NemuIpcError,
+    NemuIpcImpl,
+    NemuIpcImplConfig,
+)
 
 __all__ = [
     "ExternalRendererIpc",
+    "NemuIpcDisplayNotFoundError",
+    "NemuIpcError",
     "NemuIpcImpl",
     "NemuIpcImplConfig",
 ]